Feat: Add RISCV initial CPU and trap source
The RISCV's CPU source currently initializes the machine mode by
setting the mstatus register to 0 and enabling software interrupts
which later need to be handler appropriatly.

the init_riscv() is needed to fully initialize RISCV's CPU, this is
required as this will be later called within a higher software
stack (the kernel itself).